+++ src/game/common/system/mempoolfact.h
@@ -35,7 +35,7 @@ public:
     void Destroy_Dynamic_Memory_Allocator(DynamicMemoryAllocator *allocator);
     void Reset();
 
-    void *operator new(size_t size) throw() { return Raw_Allocate_No_Zero(size); }
+    void *operator new(size_t size) throw() { return Raw_Allocate_No_Zero(static_cast<int>(size)); }
 
     void operator delete(void *obj) { Raw_Free(obj); }
